app/requirements.txt
aiofiles==24.1.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
black==22.1.0
blue==0.9.1
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
click-spinner==0.1.10
colorama==0.4.6
croniter==1.4.1
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.1.2
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
fastapi-utilities==0.2.0
fastapi_amis_admin==0.7.2
flake8==4.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isort==5.13.2
Jinja2==3.1.4
makefun==1.15.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.6.1
mdurl==0.1.2
mypy-extensions==1.0.0
pathspec==0.12.1
platformdirs==4.2.2
psycopg2==2.9.9
pwdlib==0.2.0
pycodestyle==2.8.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyflakes==2.4.0
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlalchemy_database==0.1.2
sqlmodel==0.0.22
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uvicorn==0.30.4
watchfiles==0.22.0
websockets==12.0
